refactor: State에 Running, Finished 추상클래스 추가, 중복코드 제거
