+++ src/main/java/bowling/domain/state/Finished.java
@@ -1,0 +1,18 @@
+package bowling.domain.state;
+
+public abstract class Finished implements State {
+    @Override
+    public boolean isStrike() {
+        return false;
+    }
+
+    @Override
+    public boolean isSpare() {
+        return false;
+    }
+
+    @Override
+    public boolean hasNext() {
+        return false;
+    }
+}
