+++ src/main/java/bowling/domain/state/Spare.java
@@ -4,7 +4,7 @@ import bowling.domain.Score;
 
 import java.util.Objects;
 
-public class Spare implements State {
+public class Spare extends Finished {
     private static final String INVALID_FIRST_SCORE = "스페어의 첫번째 값은 10보다 작아야합니다.";
     private static final String INVALID_SECOND_SCORE = "스페어의 두번째 값은 10-첫번째값보다 커야합니다.";
 
@@ -31,23 +31,8 @@ public class Spare implements State {
     }
 
     @Override
-    public boolean isStrike() {
-        return false;
-    }
-
-    @Override
     public boolean isSpare() {
         return true;
-    }
-
-    @Override
-    public boolean isReady() {
-        return false;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return false;
     }
 
     @Override
