+++ src/main/java/bowling/domain/state/State.java
@@ -5,7 +5,6 @@ import bowling.domain.Score;
 public interface State {
     boolean isStrike();
     boolean isSpare();
-    boolean isReady();
     boolean hasNext();
     boolean canAccumulate();
     Score eval();
